feat: Add bare bones links page
This adds a bare bones "links" page. This is not a stopping point, but
rather something that I can build on as a template for other sorts of
miscellania. It's also not completely fleshed out, as I added exactly >1
items so that I could ensure iteration is iterating as it should be.