refactor: Replace HTML with xexpr, clean up
The primary goal of this refactoring was to replace the HTML template
with a single xexpr. To that end, this has been successful.

In addition to the above, there are some minor stylistic changes (i.e.
removing commented-out next/previous pages, as my feelings on those
buttons have changed) as well as variances in formatting between this
layout and the old layout (namely some miscellaneous whitespace
changes).

Finally, we remove the unused functions we stopped using when we removed
the relative navigation.